erdantic: rename {typing, pydantic, dataclasses}
This avoids namespace conflict and makes the cli usable without
forcing me to setup.py install for testing changes.
