+++ erdantic/__init__.py
@@ -1,6 +1,4 @@
 from erdantic.erd import create, draw, EntityRelationshipDiagram, to_dot
-import erdantic.pydantic  # noqa: F401
-import erdantic.dataclasses  # noqa: F401
 from erdantic.version import __version__
 
 __version__
